+++ src/com/atoennis/fuellog/domain/Trip.java
@@ -12,10 +12,10 @@ public class Trip
     public final int    id;
     public final Date   date;
     public final int    distance;
-    public final int    volume;
+    public final double volume;
     public final double volumePrice;
 
-    public Trip(int id, Date date, int distance, int volume, double volumePrice)
+    public Trip(int id, Date date, int distance, double volume, double volumePrice)
     {
         this.id = id;
         this.date = date;
@@ -50,7 +50,7 @@ public class Trip
                 .getColumnIndex(TripEntry.COLUMN_NAME_TRIP_VOLUME_PRICE));
 
             return new Trip(id, new Date(milliseconds), Integer.valueOf(distance).intValue(),
-                Integer.valueOf(volume).intValue(), Double.valueOf(volumePrice).doubleValue());
+                Double.valueOf(volume).doubleValue(), Double.valueOf(volumePrice).doubleValue());
         }
 
         return new Trip(0, new Date(), 0, 0, 0);
